:truck: Moved Strict PartialOrd tests to tests dir
